backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
